songcrud/musicapp/requirement.txt
django-admin startproject songcrud
python manage.py startapp musicapp
python manage.py migrate
python manage.py makemigrations musicapp
python manage.py sqlmigrate musicapp 0001 
python manage.py shell
python manage.py createsuperuser
